spectra-plot.html -> spectra-plot-old.html, spectra-plot-multi -> spectra-plot.html
