--- webapp/apps/spectra-plot-multi.html
@@ -1,44 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-  <head>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta charset="UTF-8"> 
-
-    <!-- Import Bootstrap CSS -->
-    <link rel="stylesheet" 
-        href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css">
-
-    <!-- CSS Files -->
-    <link rel="stylesheet" href="css/template.css" type="text/css">
-    <link rel="stylesheet" href="css/D3View.css" type="text/css">
-    <link rel="stylesheet" href="css/Gmm.css"  type="text/css">
-  </head>
-
-  <body>
-    <div id="content"></div> 
-
-    <!-- Import jQuery -->
-    <script 
-        src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js">
-    </script>
-
-    <!-- Import Bootstrap JavaScript -->
-    <script 
-        src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js">
-    </script>
-
-    <!-- Import D3 -->
-    <script src="https://d3js.org/d3.v4.min.js" charset="utf-8"></script>
-    
-    <!-- Run main JavaScript -->
-    <script type='module'>
-      import Config from './js/lib/Config.js';
-      import SpectraMulti from './js/SpectraMulti.js';
-      
-      Config.getConfig(SpectraMulti);
-    </script>
-
-  </body>
-
-</html>
